kube-aggregator: move openapi download where it is used
