+++ staging/src/k8s.io/kube-aggregator/pkg/controllers/openapi/downloader_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package aggregator
+package openapi
 
 import (
 	"fmt"
